+++ rpcs3/emucore.vcxproj.filters
@@ -2950,6 +2950,9 @@
     <ClInclude Include="Emu\Cell\Modules\libsvc1d.h">
       <Filter>Emu\Cell\Modules</Filter>
     </ClInclude>
+    <ClInclude Include="Emu\RSX\Common\texture_cache_blit_helpers.h">
+      <Filter>Emu\GPU\RSX\Common</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="Emu\RSX\Program\GLSLSnippets\GPUDeswizzle.glsl">
